# project/backend/document/default_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "个人贷款管理系统 API",
    "description": "个人贷款管理系统后端API文档",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/loan-applications": {
      "post": {
        "tags": [
          "贷款申请管理"
        ],
        "summary": "提交贷款申请",
        "description": "用户提交贷款申请",
        "operationId": "apply",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultString"
                }
              }
            }
          }
        }
      }
    },
    "/api/loan-applications/my/{applicationId}/withdraw": {
      "post": {
        "tags": [
          "贷款申请管理"
        ],
        "summary": "撤回申请",
        "description": "用户撤回自己的贷款申请",
        "operationId": "withdrawApplication",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "description": "申请ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/submit-all": {
      "post": {
        "tags": [
          "认证管理"
        ],
        "summary": "提交认证材料",
        "description": "用户提交所有认证材料，包括身份证、银行卡、房产证明、车辆证明、工作证明、工资证明、社保证明、征信报告",
        "operationId": "submitAllAuth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "bankCardId",
                  "idCard"
                ],
                "type": "object",
                "properties": {
                  "idCard": {
                    "type": "string",
                    "description": "身份证号"
                  },
                  "bankCardId": {
                    "type": "string",
                    "description": "银行卡号"
                  },
                  "propertyFile": {
                    "type": "string",
                    "description": "房产证明文件",
                    "format": "binary"
                  },
                  "carFile": {
                    "type": "string",
                    "description": "车辆证明文件",
                    "format": "binary"
                  },
                  "employmentFile": {
                    "type": "string",
                    "description": "工作证明文件",
                    "format": "binary"
                  },
                  "salaryFile": {
                    "type": "string",
                    "description": "工资证明文件",
                    "format": "binary"
                  },
                  "socialSecurityFile": {
                    "type": "string",
                    "description": "社保证明文件",
                    "format": "binary"
                  },
                  "creditReportFile": {
                    "type": "string",
                    "description": "征信报告文件",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "认证管理"
        ],
        "summary": "用户注册",
        "description": "用户注册新账号",
        "operationId": "userRegister",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultRegisterResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "认证管理"
        ],
        "summary": "用户登录",
        "description": "用户通过手机号和密码进行登录",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultLoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/loan-applications/{applicationId}": {
      "get": {
        "tags": [
          "贷款申请管理"
        ],
        "summary": "获取申请详情",
        "description": "管理员获取任意用户的单个贷款申请详情",
        "operationId": "getAdminApplication",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "description": "申请ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultAdminGetAppResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/loan-applications/user/{userId}": {
      "get": {
        "tags": [
          "贷款申请管理"
        ],
        "summary": "获取用户申请列表",
        "description": "管理员获取指定用户的所有贷款申请详情",
        "operationId": "getAdminAllApplications",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "用户ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListAdminGetAppResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/loan-applications/my/{applicationId}": {
      "get": {
        "tags": [
          "贷款申请管理"
        ],
        "summary": "查看单个申请",
        "description": "用户查看自己的单个贷款申请详情",
        "operationId": "userGetApplication",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "description": "申请ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultUserGetAppResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/work-cert/{workCertId}": {
      "get": {
        "tags": [
          "认证管理"
        ],
        "summary": "获取工作认证信息",
        "description": "管理员根据工作认证ID查询工作认证详细信息",
        "operationId": "getWorkCert",
        "parameters": [
          {
            "name": "workCertId",
            "in": "path",
            "description": "工作认证ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/tri-cert/{triCertId}": {
      "get": {
        "tags": [
          "认证管理"
        ],
        "summary": "获取第三方认证信息",
        "description": "管理员根据第三方认证ID查询第三方认证详细信息",
        "operationId": "getTriCert",
        "parameters": [
          {
            "name": "triCertId",
            "in": "path",
            "description": "第三方认证ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/immovables-cert/{immovableCertId}": {
      "get": {
        "tags": [
          "认证管理"
        ],
        "summary": "获取不动产认证信息",
        "description": "管理员根据不动产认证ID查询不动产认证详细信息",
        "operationId": "getImmovablesCert",
        "parameters": [
          {
            "name": "immovableCertId",
            "in": "path",
            "description": "不动产认证ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200,
            "refType": null
          },
          "data": {
            "type": "string",
            "description": "响应数据",
            "refType": null
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功",
            "refType": null
          }
        },
        "description": "API响应结果"
      },
...
      "ApiResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
...
      "ApplicationRequest": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "optionId": {
            "type": "integer",
            "format": "int64"
          },
          "term": {
            "type": "integer",
            "format": "int32"
          },
          "loanAmount": {
            "type": "number"
          }
        },
        "description": "贷款申请信息"
      },
      "RegisterRequest": {
        "required": [
          "name",
          "password",
          "phone"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 20,
            "minLength": 2,
            "type": "string",
            "description": "用户姓名",
            "example": "张三"
          },
          "phone": {
            "pattern": "^1[3-9]\\d{9}$",
            "type": "string",
            "description": "用户手机号",
            "example": "13800138000"
          },
          "password": {
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%&*()_+=])[A-Za-z\\d!@#$%&*()_+=]{8,20}$",
            "type": "string",
            "description": "用户密码",
            "example": "@Zff1234"
          }
        },
        "description": "注册请求参数"
      },
      "ApiResultRegisterResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/RegisterResponse"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "响应数据"
      },
      "LoginRequest": {
        "required": [
          "password",
          "phone"
        ],
        "type": "object",
        "properties": {
          "phone": {
            "pattern": "^1[3-9]\\d{9}$",
            "type": "string",
            "description": "用户手机号",
            "example": "13800138000"
          },
          "password": {
            "type": "string",
            "description": "用户密码",
            "example": "@Zff1234"
          }
        },
        "description": "登录请求参数"
      },
      "ApiResultLoginResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/LoginResponse"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT令牌",
            "example": "eyJhbGciOiJIUzI1NiJ9..."
          }
        },
        "description": "登录响应"
      },
...
      "AdminGetAppResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "userName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "loanAmount": {
            "type": "number"
          },
          "interestRate": {
            "type": "number"
          },
          "loanPeriod": {
            "type": "integer",
            "format": "int32"
          },
          "term": {
            "type": "integer",
            "format": "int32"
          },
          "repaidType": {
            "type": "string",
            "enum": [
              "等额本息",
              "等额本金",
              "先息后本",
              "一次性还本付息"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "审核中",
              "已通过",
              "AI拒绝",
              "人工拒绝",
              "已取消"
            ]
          },
          "applyTime": {
            "type": "string",
            "format": "date-time"
          },
          "reviewTime": {
            "type": "string",
            "format": "date-time"
          },
          "rejectReason": {
            "type": "string"
          }
        },
        "description": "响应数据"
      },
      "ApiResultAdminGetAppResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminGetAppResponse"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "ApiResultListAdminGetAppResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "type": "array",
            "description": "响应数据",
            "items": {
              "$ref": "#/components/schemas/AdminGetAppResponse"
            }
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
...
      "UserGetAppResponse": {
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "integer",
            "format": "int64"
          },
          "productName": {
            "type": "string"
          },
          "loanAmount": {
            "type": "number"
          },
          "interestRate": {
            "type": "number"
          },
          "loanPeriod": {
            "type": "integer",
            "format": "int32"
          },
          "term": {
            "type": "integer",
            "format": "int32"
          },
          "repaidType": {
            "type": "string",
            "enum": [
              "等额本息",
              "等额本金",
              "先息后本",
              "一次性还本付息"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "审核中",
              "已通过",
              "AI拒绝",
              "人工拒绝",
              "已取消"
            ]
          },
          "applyTime": {
            "type": "string",
            "format": "date-time"
          },
          "reviewTime": {
            "type": "string",
            "format": "date-time"
          },
          "rejectReason": {
            "type": "string"
          }
        },
        "description": "响应数据"
      },
      "ApiResultUserGetAppResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/UserGetAppResponse"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "ApiResultObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
...
    }
  }
}